+++ pytket/binders/passes.cpp
@@ -752,9 +752,12 @@ PYBIND11_MODULE(passes, m) {
       "FlattenRelabelRegistersPass", &gen_flatten_relabel_registers_pass,
       "Removes empty Quantum wires from the Circuit and relabels all Qubit to "
       "a register from passed name. \n\n:param label: Name to relabel "
-      "remaining Qubit to, default 'q'.\n:return: A pass that removes empty "
+      "remaining Qubit to, default 'q'.\n:param relabel_classical_expressions: "
+      "Whether to relabel arguments of expressions held in `ClassicalExpBox`. "
+      "\n:return: A pass that removes empty "
       "wires and relabels.",
-      py::arg("label") = q_default_reg());
+      py::arg("label") = q_default_reg(),
+      py::arg("relabel_classical_expressions") = true);
 
   m.def(
       "RenameQubitsPass", &gen_rename_qubits_pass,
@@ -957,10 +960,14 @@ PYBIND11_MODULE(passes, m) {
       "\n:param allow_zzphase: If set to True, allows the algorithm to "
       "implement 2-qubit rotations using ZZPhase gates when deemed "
       "optimal. Defaults to False."
+      "\n:param timeout: Sets maximum out of time spent finding solution."
+      "\n:param only_reduce: Only returns modified circuit if it has "
+      "fewer two-qubit gates."
       "\n:return: a pass to perform the simplification",
       py::arg("discount_rate") = 0.7, py::arg("depth_weight") = 0.3,
       py::arg("max_lookahead") = 500, py::arg("max_tqe_candidates") = 500,
-      py::arg("seed") = 0, py::arg("allow_zzphase") = false);
+      py::arg("seed") = 0, py::arg("allow_zzphase") = false,
+      py::arg("timeout") = 100, py::arg("only_reduce") = false);
   m.def(
       "PauliSquash", &PauliSquash,
       "Applies :py:meth:`PauliSimp` followed by "
@@ -1033,7 +1040,7 @@ PYBIND11_MODULE(passes, m) {
 
   m.def(
       "CnXPairwiseDecomposition", &CnXPairwiseDecomposition,
-      "Decompose CnX gates to 2-qubit gates and single qubit gates. "
+      "Decompose CnX gates to 2-qubit gates `fand single qubit gates. "
       "For every two CnX gates, reorder their control qubits to improve "
       "the chance of gate cancellation");