+++ pytket/conanfile.py
@@ -1,5 +1,5 @@
 from conan import ConanFile
-from conan.tools.cmake import CMakeToolchain, CMake, cmake_layout, CMakeDeps
+from conan.tools.cmake import CMake, CMakeDeps, CMakeToolchain, cmake_layout
 
 
 class pytketRecipe(ConanFile):
@@ -36,9 +36,9 @@ class pytketRecipe(ConanFile):
         self.requires("nlohmann_json/3.11.3")
         self.requires("pybind11/2.13.6")
         self.requires("pybind11_json/0.2.14")
-        self.requires("symengine/0.12.0")
+        self.requires("symengine/0.13.0")
         self.requires("tkassert/0.3.4@tket/stable")
-        self.requires("tket/1.3.39@tket/stable")
+        self.requires("tket/1.3.46@tket/stable")
         self.requires("tklog/0.3.3@tket/stable")
         self.requires("tkrng/0.3.3@tket/stable")
         self.requires("tktokenswap/0.3.9@tket/stable")